cloud_function/requirements.txt
functions-framework==3.*
firebase-admin
google-cloud-firestore
google-cloud-storage